requirements.txt
apscheduler
PyYAML
notifiers
requests
beautifulsoup4